Device memory allocation and a modular C API. - Added host-mapped and device-only `Malloc` / `Free` (pooled and   dedicated), plus wait-free `BumpAllocator` and O(1) `OffsetAllocator`. - Added Vulkan memory backing (VMA) and vendored   `ankerl::unordered_dense` for allocation storage. - Added `Version`, `HeaderVersion` / `LinkedVersion`, and   `Error::VersionMismatch` when headers and the linked binary disagree. - Added `ShaderFormat`, shader magics, `ShaderBlob`, and   `NativeShaderFormat` so `CreateProgram` takes a borrowed native blob. - Added `DualRange`, `GpuPtr` / `DualPtr` arithmetic, `SizeAlign`,   `CopyGranularity`, and built-in `SamplerSlot`s. Moved opaque   descriptors to the Vulkan backend. - Added adapter and `DeviceInfo` fields for texture-heap alignment,   timestamp period, copy granularity, and per-queue timestamps.   `CommandPoolDesc::max_timestamps` budgets timestamp writes. - Changed validation to compile-in support   (`APERTURE_ENABLE_VALIDATION_SUPPORT`) and enable it per instance   (`InstanceDesc::enable_validation`). - Widened `Capability` to `uint64_t`, `Hazard` and `TextureUsage` to   `uint32_t`. Removed `ArenaOverflow`, `Program`, and `MatrixLayout`. - Expanded `vector` / `matrix` (element access, row/column get/set, full   2–4 aliases). `log::GetCustomLogger` returns the active callback. - Split C bindings into `core`, `vulkan`, `cpu_root`, and `imgui`   modules (`aperture::c` is glue). Default C bindings on for top-level   builds. Added a C `basic` example. - Stopped fetching and linking Slang from the library. Install exports C   and C++ headers under `include/aperture` and keeps third-party   includes language-gated.

+++ src/core/include/aperture/capability.hpp
@@ -7,23 +7,24 @@
 namespace aperture {
 
 /// @brief Optional device features negotiated at `CreateDevice`.
-/// @details Reported on `Adapter::capabilities`.
-enum class Capability : uint16_t {
-  None = 0,
-  MeshShading = 1U << 0U,
-  RayTracingPipeline = 1U << 1U,
-  RayQuery = 1U << 2U,
-  SeparateBlend = 1U << 3U,
-  HostImageCopy = 1U << 4U,
-  UnifiedImageLayouts = 1U << 5U,
-  DeviceGeneratedCommands = 1U << 6U,
-  AsyncCompute = 1U << 7U,
-  AsyncCopy = 1U << 8U,
-  PresentFromCompute = 1U << 9U,
-  CooperativeMatrix = 1U << 10U,
-  FramebufferFetch = 1U << 11U,
-  BufferInt64Atomics = 1U << 12U,
-  SplitBarriers = 1U << 13U,
+/// @details Reported on `Adapter::capabilities`. Stored as `uint64_t` so new
+/// bits can be added without changing the type.
+enum class Capability : uint64_t {
+  None = 0ULL,
+  MeshShading = 1ULL << 0ULL,
+  RayTracingPipeline = 1ULL << 1ULL,
+  RayQuery = 1ULL << 2ULL,
+  SeparateBlend = 1ULL << 3ULL,
+  HostImageCopy = 1ULL << 4ULL,
+  UnifiedImageLayouts = 1ULL << 5ULL,
+  DeviceGeneratedCommands = 1ULL << 6ULL,
+  AsyncCompute = 1ULL << 7ULL,
+  AsyncCopy = 1ULL << 8ULL,
+  PresentFromCompute = 1ULL << 9ULL,
+  CooperativeMatrix = 1ULL << 10ULL,
+  FramebufferFetch = 1ULL << 11ULL,
+  BufferInt64Atomics = 1ULL << 12ULL,
+  SplitBarriers = 1ULL << 13ULL,
 };
 
 /// @brief Name of a `Capability` enumerator.
